+++ R/assay.R
@@ -588,7 +588,7 @@ LayerData.Assay <- function(
     stop('features are not found')
   }
   # Pull the matrix for the cells/features requested
-  return(methods::slot(object = object, name = layer)[features, cells])
+  return(methods::slot(object = object, name = layer)[features, cells, drop = FALSE])
 }
 
 #' @rdname Layers
@@ -1488,13 +1488,7 @@ subset.Assay <- function(x, cells = NULL, features = NULL, ...) {
     new(Class = 'matrix')
   }
   VariableFeatures(object = x) <- VariableFeatures(object = x)[VariableFeatures(object = x) %in% features]
-
-  # Subset feature-level metadata.
-  meta.features <- x[[]]
-  meta.features <- meta.features[features, , drop = FALSE]
-  meta.features <- droplevels(meta.features)
-  slot(object = x, name = 'meta.features') <- meta.features
-
+  slot(object = x, name = 'meta.features') <- x[[]][features, , drop = FALSE]
   validObject(object = x)
   return(x)
 }